moving logs explore to under logs page
