+++ src/pages/logspage/LogsExplorer/style.scss
